Adding HomeController, making more button show selected name.
