+++ SpeedRunApp.UI/src/components/ButtonDropdownVue.vue
@@ -1,6 +1,6 @@
 ﻿<template>
     <div class="btn-dropdown dropdown" :class="{ show : state }" @click="toggleDropdown">
-        <button class="btn dropdown-toggle" :class="btnclasses" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" ref="btn">
+        <button class="btn dropdown-toggle nowrap-elipsis" :class="btnclasses" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" ref="btn">
             <slot name="text"></slot>
         </button>
         <div class="dropdown-menu" :class="[listclasses]" aria-labelledby="dropdownMenuButton" ref="list" :style="[ state ? { display:'block' } : { display:'none' } ] ">
