+++ SpeedRunApp.UI/src/components/SpeedRunGridTabVue.vue
@@ -6,7 +6,7 @@
             </div>
         </div>
     </div> 
-    <div v-else id="divSpeedRunGridTabContainer">
+    <div v-else id="divSpeedRunGridTabContainer" style="overflow-x:hidden;">
         <div v-if="!isgame" class="row no-gutters pr-1 pt-1 pb-0">
             <div class="col tab-list">
                 <ul class="nav nav-pills">
@@ -202,6 +202,9 @@
                 levelID: '',
                 subCategoryVariableValueIDs: {},
                 variableIndex: 0,
+                // gameMoreTxt: 'More...',
+                // categoryMoreTxt: 'More...',
+                // levelMoreTxt: 'More...',
                 hideEmpty: true,
                 showCharts: true,
                 showAllData: false,
@@ -389,18 +392,19 @@
             //        });    
             //    });
             //},
-            onTabClick: function (event) {
+            onTabClick: function (event) {                
                 var type = event.target.getAttribute('data-type');
                 var value = event.target.getAttribute('data-value');
+                var name = event.target.innerHTML;
                 var variableName = event.target.getAttribute('data-variable');
-                var isMore = event.target.parentElement.classList.contains('dropdown-item');
+                var isMore = event.target.classList.contains('dropdown-item');
 
                 switch(type) {
                     case 'game':
                         this.gameID = value;
                         break;
                     case 'categoryType':
-                        this.categoryTypeID = value;
+                        this.categoryTypeID = value;                        
                         break;
                     case 'category':
                         this.categoryID = value;
@@ -455,7 +459,7 @@
                         tabitems[i].style.left = "-10000px";
                         tabitems[i].classList.remove('d-none');
                         totalWidth += tabitems[i].offsetWidth;
-                        if (totalWidth > (rows[g].offsetWidth - 100)) {
+                        if (totalWidth > ((rows[g].offsetWidth - tabitems[i].offsetWidth) - 30)) {
                             tabitems[i].classList.add('d-none');
                             moreItems[i].classList.remove('d-none');
                         } else {
@@ -464,12 +468,20 @@
                         }
                     }
 
-                    morediv.style.display = morediv.querySelectorAll('a:not(.d-none)').length > 0 ? 'block' : 'none';
-                    
-                    if (morediv.querySelectorAll('a:not(.d-none).active').length > 0) {
-                        morebtn.classList.add('active');
+                    var items = Array.from(morediv.querySelectorAll('a:not(.d-none)'));
+                    if (items.length > 0) {
+                        var item = items.find(i=>i.classList.contains('active'));
+                        if (item) {
+                            morebtn.classList.add('active');
+                        } else {
+                            item = items[0];
+                            morebtn.classList.remove('active');
+                        }
+
+                        morebtn.innerHTML = item.innerHTML;
+                        morediv.style.display = 'block';
                     } else {
-                        morebtn.classList.remove('active');
+                        morediv.style.display = 'none';                       
                     }
 
                     var ww = document.documentElement.clientWidth;
